fix: unread items sometimes appear as read
This bug was caused by items that did not have voting. The id was being
fetched from the voting buttons, wich caused the indexOf to be called
with and empty string, returning positive.

This commit also refactors all the parsing code to be simpler to read
and increases performance by only attaching one event listener, instead
of 30.